+++ lib/squib/conf.rb
@@ -10,9 +10,9 @@ module Squib
     DEFAULTS = {
       'antialias'     => 'best',
       'backend'       => 'memory',
-      'count_format'  => SYSTEM_DEFAULTS[:count_format],
+      'count_format'  => '%02d',
       'custom_colors' => {},
-      'dir'           => SYSTEM_DEFAULTS[:dir],
+      'dir'           => '_output',
       'hint'          => :none,
       'img_dir'       => '.',
       'progress_bars' => false,
@@ -52,6 +52,7 @@ module Squib
         Squib::logger.info { "  using config: #{file}" }
         yaml = YAML.load_file(file) || {}
       end
+      warn_unrecognized(yaml)
       Conf.new(DEFAULTS.merge(yaml))
     end
 
@@ -111,6 +112,14 @@ module Squib
 
     def normalize_antialias
       @config_hash['antialias'] = ANTIALIAS_OPTS[@config_hash['antialias'].downcase.strip]
+    end
+
+    # Were there any unrecognized options in the config file?
+    def self.warn_unrecognized(yaml)
+      unrec = yaml.keys - DEFAULTS.keys
+      if unrec.any?
+        Squib::logger.warn "Unrecognized configuration option(s): #{unrec.join(',')}"
+      end
     end
 
   end